Replace CaretakerService with plain functions
